clean up audit model creation
