+++ src/models/auditModelGenerator.js
@@ -71,11 +71,116 @@ const removeFromAuditedTransactions = (options) => {
   }
 };
 
-const generateAuditModel = (sequelize, model) => {
-  const auditModelName = `ZAL${model.name}`;
+const generateModelClass = (sequelize, name, schema) => {
+  const auditModelName = name;
   const auditModel = class extends Model {};
+  auditModel.init(schema, {
+    sequelize,
+    modelName: auditModelName,
+    createdAt: false,
+    updatedAt: false,
+  });
+  module.exports[auditModelName] = auditModel;
+  return auditModel;
+};
 
-  auditModel.init({
+const generateZALDDL = (sequelize) => {
+  const name = 'ZALDDL';
+  const schema = {
+    id: {
+      type: DataTypes.BIGINT,
+      allowNull: false,
+      defaultValue: null,
+      primaryKey: true,
+      autoIncrement: true,
+    },
+    ddl_timestamp: {
+      allowNull: false,
+      type: DataTypes.DATE,
+    },
+    ddl_by: {
+      type: DataTypes.INTEGER,
+      allowNull: true,
+      defaultValue: null,
+      comment: null,
+    },
+    session_sig: {
+      type: DataTypes.TEXT,
+      allowNull: true,
+      defaultValue: null,
+    },
+    ddl_txid: {
+      type: DataTypes.UUID,
+      allowNull: false,
+      validate: { isUUID: 'all' },
+    },
+    descriptor_id: {
+      type: DataTypes.INTEGER,
+      allowNull: true,
+      defaultValue: null,
+    },
+    command_tag: {
+      type: DataTypes.STRING,
+    },
+    object_type: {
+      type: DataTypes.STRING,
+    },
+    schema_name: {
+      type: DataTypes.STRING,
+    },
+    object_identity: {
+      type: DataTypes.STRING,
+    },
+  };
+
+  return generateModelClass(sequelize, name, schema);
+};
+
+const generateZADescriptor = (sequelize) => {
+  const name = 'ZADescriptor';
+  const schema = {
+    id: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+      defaultValue: null,
+      primaryKey: true,
+      autoIncrement: true,
+    },
+    descriptor: {
+      allowNull: false,
+      type: DataTypes.TEXT,
+    },
+  };
+
+  return generateModelClass(sequelize, name, schema);
+};
+
+const generateZAFilter = (sequelize) => {
+  const name = 'ZAFilter';
+  const schema = {
+    id: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+      defaultValue: null,
+      primaryKey: true,
+      autoIncrement: true,
+    },
+    tableName: {
+      allowNull: true,
+      type: DataTypes.STRING,
+    },
+    columnName: {
+      allowNull: false,
+      type: DataTypes.STRING,
+    },
+  };
+
+  return generateModelClass(sequelize, name, schema);
+};
+
+const generateAuditModel = (sequelize, model) => {
+  const name = `ZAL${model.name}`;
+  const schema = {
     id: {
       type: DataTypes.BIGINT,
       allowNull: false,
@@ -129,14 +234,9 @@ const generateAuditModel = (sequelize, model) => {
       allowNull: true,
       defaultValue: null,
     },
-  }, {
-    sequelize,
-    modelName: auditModelName,
-    createdAt: false,
-    updatedAt: false,
-  });
-  module.exports[auditModelName] = auditModel;
-  return auditModel;
+  };
+
+  return generateModelClass(sequelize, name, schema);
 };
 
 // eslint-disable-next-line
@@ -201,6 +301,9 @@ const attachHooksForAuditing = (sequelize) => {
 };
 
 export {
+  generateZALDDL,
+  generateZADescriptor,
+  generateZAFilter,
   generateAuditModel,
   attachHooksForAuditing,
   addAuditTransactionSettings,
